+++ test/client/calendar/CalendarModelTest.js
@@ -35,6 +35,7 @@ import {createCalendarGroupRoot} from "../../../src/api/entities/tutanota/Calend
 import {_loadEntity} from "../../../src/api/common/EntityFunctions"
 import {NotFoundError} from "../../../src/api/common/error/RestError"
 import type {LoginController} from "../../../src/api/main/LoginController"
+import {ProgressTracker} from "../../../src/api/main/ProgressTracker"
 
 o.spec("CalendarModel", function () {
 	o.spec("addDaysForEvent", function () {
@@ -722,7 +723,7 @@ o.spec("CalendarModel", function () {
 				getEventByUid: (loadUid) => uid === loadUid ? Promise.resolve(existingEvent) : Promise.resolve(null),
 				updateCalendarEvent: o.spy(() => Promise.resolve()),
 			})
-			const model = new CalendarModelImpl(makeNotifications(), eventController, workerClient, makeLoginController())
+			const model = new CalendarModelImpl(makeNotifications(), eventController, workerClient, makeLoginController(), makeProgressTracker())
 
 			await model.processCalendarUpdate("sender@example.com", {
 				method: CalendarMethod.REPLY,
@@ -767,7 +768,7 @@ o.spec("CalendarModel", function () {
 			}))
 			workerMock.eventByUid.set(uid, existingEvent)
 			const workerClient = makeWorkerClient(workerMock)
-			const model = new CalendarModelImpl(makeNotifications(), eventController, workerClient, loginController)
+			const model = new CalendarModelImpl(makeNotifications(), eventController, workerClient, loginController, makeProgressTracker())
 
 			await model.processCalendarUpdate(sender, {
 				summary: "v2", // should be ignored
@@ -813,7 +814,7 @@ o.spec("CalendarModel", function () {
 			const sender = "sender@example.com"
 			const workerMock = new WorkerMock()
 			const workerClient = makeWorkerClient(workerMock)
-			const model = new CalendarModelImpl(makeNotifications(), eventController, workerClient, loginController)
+			const model = new CalendarModelImpl(makeNotifications(), eventController, workerClient, loginController, makeProgressTracker())
 
 			await model.processCalendarUpdate(sender, {
 				summary: "1",
@@ -858,7 +859,7 @@ o.spec("CalendarModel", function () {
 			workerMock.eventByUid.set(uid, existingEvent)
 
 			const workerClient = makeWorkerClient(workerMock)
-			const model = new CalendarModelImpl(makeNotifications(), eventController, workerClient, loginController)
+			const model = new CalendarModelImpl(makeNotifications(), eventController, workerClient, loginController, makeProgressTracker())
 
 			const sentEvent = createCalendarEvent({
 				summary: "v2",
@@ -901,7 +902,7 @@ o.spec("CalendarModel", function () {
 			workerMock.eventByUid.set(uid, existingEvent)
 
 			const workerClient = makeWorkerClient(workerMock)
-			const model = new CalendarModelImpl(makeNotifications(), eventController, workerClient, loginController)
+			const model = new CalendarModelImpl(makeNotifications(), eventController, workerClient, loginController, makeProgressTracker())
 
 			const sentEvent = createCalendarEvent({
 				summary: "v2",
@@ -942,7 +943,7 @@ o.spec("CalendarModel", function () {
 				workerMock.eventByUid.set(uid, existingEvent)
 
 				const workerClient = makeWorkerClient(workerMock)
-				const model = new CalendarModelImpl(makeNotifications(), eventController, workerClient, loginController)
+				const model = new CalendarModelImpl(makeNotifications(), eventController, workerClient, loginController, makeProgressTracker())
 
 				const sentEvent = createCalendarEvent({uid, sequence: "2", organizer: createEncryptedMailAddress({address: sender})})
 				await model.processCalendarUpdate(sender, {
@@ -970,7 +971,7 @@ o.spec("CalendarModel", function () {
 				workerMock.eventByUid.set(uid, existingEvent)
 
 				const workerClient = makeWorkerClient(workerMock)
-				const model = new CalendarModelImpl(makeNotifications(), eventController, workerClient, loginController)
+				const model = new CalendarModelImpl(makeNotifications(), eventController, workerClient, loginController, makeProgressTracker())
 
 				const sentEvent = createCalendarEvent({uid, sequence: "2", organizer: createEncryptedMailAddress({address: sender})})
 				await model.processCalendarUpdate("another-sender", {
@@ -986,6 +987,12 @@ o.spec("CalendarModel", function () {
 
 function makeNotifications(): Notifications {
 	return downcast({})
+}
+
+function makeProgressTracker(): ProgressTracker {
+	return downcast({
+		register: () => 0
+	})
 }
 
 function makeEventController(): {eventController: EventController, sendEvent: (EntityUpdateData) => void} {
